+++ helpers/utils.h
@@ -9,7 +9,7 @@ using std::string;
 namespace Utils {
   unsigned int get_random_x(unsigned int factor);
   unsigned int get_random_y(unsigned int factor);
-  double random(int start, int end);
+  double random(double start, double end);
   unsigned int random_int(int start, int end);
   bool passed_debounce_time(int last_tick);
   void append_to_file(string filename, string data);
